BIG UPDATE Rework character controller and make enemies move
In order to accommodate clean bouncing and predictable movement for the players and enemies, movement has been switched to direct control of the transform rather than the velocity of the rigidbody.

Enemies now travel without collisions with eachother or the player

Enemies travel at a randomly assigned constant velocity and bounce off the edges of the screen

Player bounces off edges, with a minimum bounce amount and hard coded edge limits preventing moving off the screen.

Bounce sound effect added (50% volume for enemies, 100% for player)

Adaptive screen shake added to player bouncing, adjusts to the speed of impact.